cherry pick (ish) refactoring and fixes from master 🍒
In order to cherry pick in the fixes from #2393 and #2471, we first need
to apply the refactoring/changes these were built on in
https://github.com/tektoncd/pipeline/commit/a8946d42a91808b96047a1a3090237e787d5d395#diff-e2b4c96e5145aa139a97f864d9f9f2a1
but unfortunately this both refactored/changed the code, AND added
pipeline results. Since we don't want to add a new feature in a patch
release, this commit extracts the refactoring (and fix?) from that
commit.
