+++ pkg/apis/pipeline/v1alpha1/pipeline_types.go
@@ -164,9 +164,12 @@ func (pt PipelineTask) Deps() []string {
 	}
 	// Add any dependents from task results
 	for _, param := range pt.Params {
-		if resultRefs, err := v1beta1.NewResultRefs(param); err == nil {
-			for _, resultRef := range resultRefs {
-				deps = append(deps, resultRef.PipelineTask)
+		expressions, ok := v1beta1.GetVarSubstitutionExpressionsForParam(param)
+		if ok {
+			if resultRefs, err := v1beta1.NewResultRefs(expressions); err == nil {
+				for _, resultRef := range resultRefs {
+					deps = append(deps, resultRef.PipelineTask)
+				}
 			}
 		}
 	}
