+++ pkg/apis/pipeline/v1beta1/param_types_test.go
@@ -20,7 +20,6 @@ import (
 	"context"
 	"encoding/json"
 	"reflect"
-	"sort"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
@@ -186,370 +185,5 @@ func TestArrayOrString_MarshalJSON(t *testing.T) {
 		if string(result) != c.result {
 			t.Errorf("Failed to marshal input '%v': expected: %+v, got %q", input, c.result, string(result))
 		}
-	}
-}
-
-func TestNewResultReference(t *testing.T) {
-	type args struct {
-		param v1beta1.Param
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    []*v1beta1.ResultRef
-		wantErr bool
-	}{
-		{
-			name: "Test valid expression",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(tasks.sumTask.results.sumResult)",
-					},
-				},
-			},
-			want: []*v1beta1.ResultRef{
-				{
-					PipelineTask: "sumTask",
-					Result:       "sumResult",
-				},
-			},
-			wantErr: false,
-		}, {
-			name: "Test valid expression: substitution within string",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "sum-will-go-here -> $(tasks.sumTask.results.sumResult)",
-					},
-				},
-			},
-			want: []*v1beta1.ResultRef{
-				{
-					PipelineTask: "sumTask",
-					Result:       "sumResult",
-				},
-			},
-			wantErr: false,
-		}, {
-			name: "Test valid expression: multiple substitution",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(tasks.sumTask1.results.sumResult) and another $(tasks.sumTask2.results.sumResult)",
-					},
-				},
-			},
-			want: []*v1beta1.ResultRef{
-				{
-					PipelineTask: "sumTask1",
-					Result:       "sumResult",
-				}, {
-					PipelineTask: "sumTask2",
-					Result:       "sumResult",
-				},
-			},
-			wantErr: false,
-		}, {
-			name: "Test invalid expression: first separator typo",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(task.sumTasks.results.sumResult)",
-					},
-				},
-			},
-			want:    nil,
-			wantErr: true,
-		}, {
-			name: "Test invalid expression: third separator typo",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(tasks.sumTasks.result.sumResult)",
-					},
-				},
-			},
-			want:    nil,
-			wantErr: true,
-		}, {
-			name: "Test invalid expression: param substitution shouldn't be considered result ref",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(params.paramName)",
-					},
-				},
-			},
-			want:    nil,
-			wantErr: true,
-		}, {
-			name: "Test invalid expression: One bad and good result substitution",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "good -> $(tasks.sumTask1.results.sumResult) bad-> $(task.sumTask2.results.sumResult)",
-					},
-				},
-			},
-			want:    nil,
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := v1beta1.NewResultRefs(tt.args.param)
-			if tt.wantErr != (err != nil) {
-				t.Errorf("TestNewResultReference/%s wantErr %v, error = %v", tt.name, tt.wantErr, err)
-				return
-			}
-			if d := cmp.Diff(tt.want, got); d != "" {
-				t.Errorf("TestNewResultReference/%s (-want, +got) = %v", tt.name, d)
-			}
-		})
-	}
-}
-
-func TestHasResultReference(t *testing.T) {
-	type args struct {
-		param v1beta1.Param
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantRef []*v1beta1.ResultRef
-	}{
-		{
-			name: "Test valid expression",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(tasks.sumTask.results.sumResult)",
-					},
-				},
-			},
-			wantRef: []*v1beta1.ResultRef{
-				{
-					PipelineTask: "sumTask",
-					Result:       "sumResult",
-				},
-			},
-		}, {
-			name: "Test valid expression with dashes",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(tasks.sum-task.results.sum-result)",
-					},
-				},
-			},
-			wantRef: []*v1beta1.ResultRef{
-				{
-					PipelineTask: "sum-task",
-					Result:       "sum-result",
-				},
-			},
-		}, {
-			name: "Test invalid expression: param substitution shouldn't be considered result ref",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(params.paramName)",
-					},
-				},
-			},
-			wantRef: nil,
-		}, {
-			name: "Test valid expression in array",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:     v1beta1.ParamTypeArray,
-						ArrayVal: []string{"$(tasks.sumTask.results.sumResult)", "$(tasks.sumTask2.results.sumResult2)"},
-					},
-				},
-			},
-			wantRef: []*v1beta1.ResultRef{
-				{
-					PipelineTask: "sumTask",
-					Result:       "sumResult",
-				},
-				{
-					PipelineTask: "sumTask2",
-					Result:       "sumResult2",
-				},
-			},
-		}, {
-			name: "Test valid expression in array - no ref in first element",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:     v1beta1.ParamTypeArray,
-						ArrayVal: []string{"1", "$(tasks.sumTask2.results.sumResult2)"},
-					},
-				},
-			},
-			wantRef: []*v1beta1.ResultRef{
-				{
-					PipelineTask: "sumTask2",
-					Result:       "sumResult2",
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, _ := v1beta1.NewResultRefs(tt.args.param)
-			sort.Slice(got, func(i, j int) bool {
-				if got[i].PipelineTask > got[j].PipelineTask {
-					return false
-				}
-				if got[i].Result > got[j].Result {
-					return false
-				}
-				return true
-			})
-			if d := cmp.Diff(tt.wantRef, got); d != "" {
-				t.Errorf("TestHasResultReference/%s (-want, +got) = %v", tt.name, d)
-			}
-		})
-	}
-}
-
-func TestLooksLikeResultRef(t *testing.T) {
-	type args struct {
-		param v1beta1.Param
-	}
-	tests := []struct {
-		name string
-		args args
-		want bool
-	}{
-		{
-			name: "test expression that is a result ref",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(tasks.sumTasks.results.sumResult)",
-					},
-				},
-			},
-			want: true,
-		}, {
-			name: "test expression: looks like result ref, but typo in 'task' separator",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(task.sumTasks.results.sumResult)",
-					},
-				},
-			},
-			want: true,
-		}, {
-			name: "test expression: looks like result ref, but typo in 'results' separator",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(tasks.sumTasks.result.sumResult)",
-					},
-				},
-			},
-			want: true,
-		}, {
-			name: "test expression: missing 'task' separator",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(sumTasks.results.sumResult)",
-					},
-				},
-			},
-			want: false,
-		}, {
-			name: "test expression: missing variable substitution",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "tasks.sumTasks.results.sumResult",
-					},
-				},
-			},
-			want: false,
-		}, {
-			name: "test expression: param substitution shouldn't be considered result ref",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(params.someParam)",
-					},
-				},
-			},
-			want: false,
-		}, {
-			name: "test expression: one good ref, one bad one should return true",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:      v1beta1.ParamTypeString,
-						StringVal: "$(tasks.sumTasks.results.sumResult) $(task.sumTasks.results.sumResult)",
-					},
-				},
-			},
-			want: true,
-		}, {
-			name: "test expression: inside array parameter",
-			args: args{
-				param: v1beta1.Param{
-					Name: "param",
-					Value: v1beta1.ArrayOrString{
-						Type:     v1beta1.ParamTypeArray,
-						ArrayVal: []string{"$(tasks.sumTask.results.sumResult)", "$(tasks.sumTask2.results.sumResult2)"},
-					},
-				},
-			},
-			want: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := v1beta1.LooksLikeContainsResultRefs(tt.args.param); got != tt.want {
-				t.Errorf("LooksLikeContainsResultRefs() = %v, want %v", got, tt.want)
-			}
-		})
 	}
 }
