+++ pkg/reconciler/pipelinerun/resources/resultrefresolution.go
@@ -48,10 +48,18 @@ func ResolveResultRefs(pipelineRunState PipelineRunState, targets PipelineRunSta
 	return removeDup(allResolvedResultRefs), nil
 }
 
-// extractResultRefsFromParam resolves any ResultReference that are found in param
+// extractResultRefsForParam resolves any ResultReference that are found in param
 // Returns nil if none are found
-func extractResultRefsFromParam(pipelineRunState PipelineRunState, param v1beta1.Param) (ResolvedResultRefs, error) {
-	if resultRefs, err := v1beta1.NewResultRefs(param); err == nil {
+func extractResultRefsForParam(pipelineRunState PipelineRunState, param v1beta1.Param) (ResolvedResultRefs, error) {
+	expressions, ok := v1beta1.GetVarSubstitutionExpressionsForParam(param)
+	if ok {
+		return extractResultRefs(expressions, pipelineRunState)
+	}
+	return nil, nil
+}
+
+func extractResultRefs(expressions []string, pipelineRunState PipelineRunState) (ResolvedResultRefs, error) {
+	if resultRefs, err := v1beta1.NewResultRefs(expressions); err == nil {
 		var resolvedResultRefs ResolvedResultRefs
 		for _, resultRef := range resultRefs {
 			resolvedResultRef, err := resolveResultRef(pipelineRunState, resultRef)
@@ -85,7 +93,7 @@ func removeDup(refs ResolvedResultRefs) ResolvedResultRefs {
 func convertParamsToResultRefs(pipelineRunState PipelineRunState, target *ResolvedPipelineRunTask) (ResolvedResultRefs, error) {
 	var resolvedParams ResolvedResultRefs
 	for _, param := range target.PipelineTask.Params {
-		resolvedResultRefs, err := extractResultRefsFromParam(pipelineRunState, param)
+		resolvedResultRefs, err := extractResultRefsForParam(pipelineRunState, param)
 		if err != nil {
 			return nil, fmt.Errorf("unable to find result referenced by param %q in pipeline task %q: %w", param.Name, target.PipelineTask.Name, err)
 		}
